Split file.go and util.go off from main.go
To keep a clear separation between the directory generator and the
file generator.
