+++ main.go
@@ -2,41 +2,7 @@ package main
 
 import (
 	"flag"
-	"fmt"
-	"io"
-	"os"
-	"strings"
 )
-
-var templateStart = `package %s
-var %s = ` + "`"
-
-func convertName(name string) string {
-	name = strings.Replace(name, "/", "_", -1)
-	name = strings.Replace(name, ".", "_", -1)
-	name = strings.Replace(name, "-", "_", -1)
-	return strings.Title(name)
-}
-
-func ConvertFile(in, out, pkg string) error {
-	infi, err := os.Open(in)
-	if err != nil {
-		return err
-	}
-
-	ofi, err := os.Create(out)
-	if err != nil {
-		return err
-	}
-
-	defer ofi.Close()
-	defer infi.Close()
-
-	fmt.Fprintf(ofi, templateStart, pkg, convertName(in))
-	io.Copy(ofi, infi)
-	ofi.Write([]byte("`\n"))
-	return nil
-}
 
 func main() {
 	input := flag.String("in", "", "specify input file")
